[fs] support hfs.ls on a bucket
Teaches `hfs.ls('gs://bucket/')` to list the files and directories at the top-level of the bucket.

In `main` that command raises because this line of `_ls_no_glob` raises:

```python3
maybe_sb_and_t, maybe_contents = await asyncio.gather(
    self._size_bytes_and_time_modified_or_none(path), ls_as_dir()
)
```

In particular, `statfile` raises a cloud-specific, esoteric error about a malformed URL or empty
object names:

```python3
async def _size_bytes_and_time_modified_or_none(self, path: str) -> Optional[Tuple[int, float]]:
    try:
        # Hadoop semantics: creation time is used if the object has no notion of last modification time.
        file_status = await self.afs.statfile(path)
        return (await file_status.size(), file_status.time_modified().timestamp())
    except FileNotFoundError:
        return None
```

I decided to add a sub-class of `FileNotFoundError` which is self-describing: `IsABucketError`.

I changed most methods to raise that error when given a bucket URL. The two interesting cases:

1. `isdir`. This raises an error but I could also see this returning `True`. A bucket is like a
   directory whose path/name is empty.

2. `isfile`. This returns False but I could also see this raising an error. This just seems
   convenient, we know the bucket is not a file so we should say so.

---

Apparently `hfs.ls` had no current tests because the globbing system doesn't work with Azure
https:// URLs. I fixed it to use `AsyncFSURL.with_new_path_component` which is resilient to Azure
https weirdness. However, I had to change `with_new_path_component` to treat an empty path in a
special way. I wanted this to hold:

```
actual = str(afs.parse_url('gs://bucket').with_new_path_component('bar'))
expected = 'gs://bucket/bar'
assert actual == expected
```

But `with_new_path_component` interacts badly with `GoogleAsyncFSURL.__str__` to return this:

```
'gs://bucket//bar'
```

+++ hail/python/hailtop/aiocloud/aiogoogle/client/storage_client.py
@@ -21,6 +21,7 @@ from hailtop.aiotools.fs import (
     FileAndDirectoryError,
     MultiPartCreate,
     UnexpectedEOFError,
+    IsABucketError,
 )
 from hailtop.aiotools import FeedableAsyncIterable, WriteBuffer
 
@@ -674,6 +675,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def open(self, url: str) -> GetObjectStream:
         bucket, name = self.get_bucket_and_name(url)
+        if name == '':
+            raise IsABucketError(url)
         return await self._storage_client.get_object(bucket, name)
 
     async def _open_from(self, url: str, start: int, *, length: Optional[int] = None) -> GetObjectStream:
@@ -686,6 +689,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def create(self, url: str, *, retry_writes: bool = True) -> WritableStream:
         bucket, name = self.get_bucket_and_name(url)
+        if name == '':
+            raise IsABucketError(url)
         params = {'uploadType': 'resumable' if retry_writes else 'media'}
         return await self._storage_client.insert_object(bucket, name, params=params)
 
@@ -706,6 +711,8 @@ class GoogleStorageAsyncFS(AsyncFS):
     async def statfile(self, url: str) -> GetObjectFileStatus:
         try:
             bucket, name = self.get_bucket_and_name(url)
+            if name == '':
+                raise IsABucketError(url)
             return GetObjectFileStatus(await self._storage_client.get_object_metadata(bucket, name), url)
         except aiohttp.ClientResponseError as e:
             if e.status == 404:
@@ -798,6 +805,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def isdir(self, url: str) -> bool:
         bucket, name = self.get_bucket_and_name(url)
+        if name == '':
+            raise IsABucketError(url)
         assert not name or name.endswith('/'), name
         params = {'prefix': name, 'delimiter': '/', 'includeTrailingDelimiter': 'true', 'maxResults': 1}
         async for page in await self._storage_client.list_objects(bucket, params=params):
@@ -808,6 +817,8 @@ class GoogleStorageAsyncFS(AsyncFS):
 
     async def remove(self, url: str) -> None:
         bucket, name = self.get_bucket_and_name(url)
+        if name == '':
+            raise IsABucketError(url)
         try:
             await self._storage_client.delete_object(bucket, name)
         except aiohttp.ClientResponseError as e:
